Move Utilities methods into a new file called DistributeUtil.java
All modularised methods are moved into DistributeUtil.java.
Change filterGender method to Void type.
